+++ history.rst
@@ -1,6 +1,9 @@
 x.x.x (xx-xx-xxxx)
 ------------------
 
+- client: strip date before parsing (#113)
+Thanks to `ndhansen <https://github.com/ndhansen>`_
+
 0.16.0 (11-03-2020)
 ------------------